+++ internal/registration/graphs/metrics_test.go
@@ -30,10 +30,11 @@ func TestGetMatchingMetrics(t *testing.T) {
 		"foo`baz": agentapi.Metric{Type: "n", Value: 0},
 		"foo`qux": agentapi.Metric{Type: "n", Value: 0},
 		"baz`qux": agentapi.Metric{Type: "n", Value: 1},
+		"baf`ding|ST[b\"YXJjaA==\":b\"eDg2XzY0\",b\"ZGlzdHJv\":b\"dWJ1bnR1LTE4LjA0\",b\"b3M=\":b\"bGludXg=\"]": agentapi.Metric{Type: "i", Value: 1},
 	}
 	noMatchDatapoint := []cosiapi.TemplateDatapoint{
 		{MetricRx: `^not_found(.+)$`}, // ensure it does not match any metric
-		{}, // e.g. a static datapoint, should be skipped and not break anything
+		{},                            // e.g. a static datapoint, should be skipped and not break anything
 	}
 
 	g, err := New(&Options{
@@ -42,7 +43,7 @@ func TestGetMatchingMetrics(t *testing.T) {
 		Config: &options.Options{
 			Host: options.Host{Name: "foo"},
 		},
-		Metrics:   &agentapi.Metrics{"test": {}},
+		Metrics:   metrics, //&agentapi.Metrics{"test": {}},
 		RegDir:    "testdata",
 		Templates: &templates.Templates{},
 	})
@@ -73,10 +74,12 @@ func TestGetMatchingMetrics(t *testing.T) {
 			t.Fatalf("unexpected error (%s)", err)
 		}
 	}
+
+	// reset back to valid metrics
+	g.metrics = metrics
+
 	// bad dp regex (syntax)
 	{
-		g.metrics = metrics
-
 		_, err := g.getMatchingMetrics([]cosiapi.TemplateDatapoint{{MetricRx: `^(bar]$`}}, emptyGloabFilters)
 		if err == nil {
 			t.Fatal("expected error")
@@ -86,8 +89,6 @@ func TestGetMatchingMetrics(t *testing.T) {
 	}
 	// bad dp regex (no subexpression)
 	{
-		g.metrics = metrics
-
 		_, err := g.getMatchingMetrics([]cosiapi.TemplateDatapoint{{MetricRx: `^bar$`}}, emptyGloabFilters)
 		if err == nil {
 			t.Fatal("expected error")
@@ -97,13 +98,11 @@ func TestGetMatchingMetrics(t *testing.T) {
 	}
 	// no matching metrics
 	{
-		g.metrics = metrics
-
 		m, err := g.getMatchingMetrics(noMatchDatapoint, emptyGloabFilters)
 		if err != nil {
 			t.Fatalf("unexpected error (%s)", err)
 		} else if len(m) > 0 {
-			t.Fatalf("unexpected zero matching metrics (%#v)", m)
+			t.Fatalf("expected zero matching metrics (%#v)", m)
 		}
 	}
 	// global include
@@ -113,13 +112,11 @@ func TestGetMatchingMetrics(t *testing.T) {
 		}
 		gf := &globalFilters{include: []*regexp.Regexp{regexp.MustCompile("bar")}} // will include 1 of 3
 
-		g.metrics = metrics
-
 		m, err := g.getMatchingMetrics(dp, gf)
 		if err != nil {
 			t.Fatalf("unexpected error (%s)", err)
 		} else if len(m) != 1 {
-			t.Fatal("unexpected 1 matching metric")
+			t.Fatal("expected 1 matching metric")
 		}
 	}
 	// global exclude
@@ -129,13 +126,11 @@ func TestGetMatchingMetrics(t *testing.T) {
 		}
 		gf := &globalFilters{exclude: []*regexp.Regexp{regexp.MustCompile("bar"), regexp.MustCompile("baz")}} // will exclude 2 of 3
 
-		g.metrics = metrics
-
 		m, err := g.getMatchingMetrics(dp, gf)
 		if err != nil {
 			t.Fatalf("unexpected error (%s)", err)
 		} else if len(m) != 1 {
-			t.Fatal("unexpected 1 matching metric")
+			t.Fatal("expected 1 matching metric")
 		}
 	}
 	// dp include
@@ -149,13 +144,11 @@ func TestGetMatchingMetrics(t *testing.T) {
 			},
 		}
 
-		g.metrics = metrics
-
 		m, err := g.getMatchingMetrics(dp, emptyGloabFilters)
 		if err != nil {
 			t.Fatalf("unexpected error (%s)", err)
 		} else if len(m) != 1 {
-			t.Fatal("unexpected 1 matching metric")
+			t.Fatal("expected 1 matching metric")
 		}
 	}
 	// dp exclude
@@ -169,13 +162,11 @@ func TestGetMatchingMetrics(t *testing.T) {
 			},
 		}
 
-		g.metrics = metrics
-
 		m, err := g.getMatchingMetrics(dp, emptyGloabFilters)
 		if err != nil {
 			t.Fatalf("unexpected error (%s)", err)
 		} else if len(m) != 1 {
-			t.Fatal("unexpected 1 matching metric")
+			t.Fatal("expected 1 matching metric")
 		}
 	}
 	// multi-match dp regex
@@ -184,13 +175,25 @@ func TestGetMatchingMetrics(t *testing.T) {
 			{MetricRx: "^foo`([^`]+)"}, // will match 3
 		}
 
-		g.metrics = metrics
-
 		m, err := g.getMatchingMetrics(dp, emptyGloabFilters)
 		if err != nil {
 			t.Fatalf("unexpected error (%s)", err)
 		} else if len(m) != 3 {
-			t.Fatal("unexpected 3 matching metrics")
+			t.Fatal("expected 3 matching metrics")
+		}
+	}
+
+	// match metric with stream tagsdp regex
+	{
+		dp := []cosiapi.TemplateDatapoint{
+			{MetricRx: "^baf`(ding|dong)"},
+		}
+
+		m, err := g.getMatchingMetrics(dp, emptyGloabFilters)
+		if err != nil {
+			t.Fatalf("unexpected error (%s)", err)
+		} else if len(m) != 1 {
+			t.Fatal("expected 1 matching metric")
 		}
 	}
 }